+++ CHANGELOG.md
@@ -5,3 +5,11 @@
 ### v1.0.1
 
 - app(update): change-added-by-to-nickname
+
+### v1.0.2
+
+- app(update): v1.0.2-fixes
+  - Added new optional description field
+  - Updated logic to not allow deletion of first restaurant
+  - Change Sunday color from red to teal
+  - UI improvements